+++ lib/repositories/mission_grace_period_repository/mission_grace_period_repository_remote.dart
@@ -1,4 +1,4 @@
-import 'package:notiyou/repositories/mission_grace_period_repository_interface.dart';
+import 'package:notiyou/repositories/mission_grace_period_repository/mission_grace_period_repository_interface.dart';
 import 'package:notiyou/repositories/supabase_table_names_constants.dart';
 import 'package:notiyou/services/supabase_service.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
