+++ lib/repositories/mission_time_repository/mission_time_repository_local.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:notiyou/models/mission_time_model.dart';
-import 'package:notiyou/repositories/mission_time_repository_interface.dart';
+import 'package:notiyou/repositories/mission_time_repository/mission_time_repository_interface.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 /// 미션 시간 데이터를 관리하는 저장소입니다.
